backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
boto3==1.35.0
openai==1.0.0
numpy==2.0.0
python-dotenv==1.0.1

# Optional (uncomment when ready to use)
# twilio==9.0.0
# pgvector==0.3.0
